requirements.txt
fastapi
uvicorn[standard]
jinja2
python-multipart
pydantic[email]
passlib[bcrypt]
python-jose[cryptography]
httpx
websockets
aiofiles
